docs : types 폴더 관리
