+++ src/types/api/auth.ts
@@ -1,0 +1,6 @@
+export interface KakaoLoginResponse {
+  userId: number;
+  profileImage: string;
+  jwtAccessToken: string;
+  jwtRefreshToken: string;
+}
