+++ src/types/api/category.ts
@@ -1,0 +1,6 @@
+export interface CategoryProps {
+  id: number;
+  categoryName: string;
+  createdAt: string;
+  updatedAt: string;
+}
