+++ src/types/api/searchHistory.ts
@@ -1,0 +1,5 @@
+export interface SearchHistoryProps {
+  id: number;
+  keyword: string;
+  searchedAt: string;
+}
